# api/openapi.yaml
openapi: 3.1.0
info:
  title: Ruammit API
  version: 0.1.0
  description: >
    Backend for the Ruammit travel social app. This is a starter contract —
    extend it module by module (auth, feed, location, chat). It doubles as the
    source for generating a Dart client for the Flutter app.
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/users/me:
    get:
      summary: Get the authenticated user's profile
      operationId: getMyProfile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The caller's profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "401": { $ref: "#/components/responses/Error" }
    patch:
      summary: Update the authenticated user's profile (setup / edit)
      description: >
        Multipart/form-data partial update. Only the fields present are changed,
        so this serves both first-time setup (after OTP) and later edits. Setting
        a display_name marks the profile complete.
      operationId: updateMyProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                display_name: { type: string, minLength: 2, maxLength: 50 }
                birth_date: { type: string, example: "1995-07-14", description: YYYY-MM-DD }
                gender: { type: string, enum: [male, female, other] }
                bio: { type: string, maxLength: 300 }
                avatar: { type: string, format: binary, description: JPG/PNG/WebP, ≤5 MB }
      responses:
        "200":
          description: The updated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Profile" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Profile:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        phone: { type: string }
        display_name: { type: string }
        avatar_url: { type: string }
        bio: { type: string }
        gender: { type: string, enum: [male, female, other] }
        birth_date: { type: string, nullable: true, description: YYYY-MM-DD }
        status: { type: string }
        profile_completed: { type: boolean }
...
    Error:
      type: object
      properties:
        error: { type: string, description: Machine-readable code }
        message: { type: string }
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
# Planned route groups (to be filled in as modules land):
#   /api/v1/auth/*       — signup, login, social, refresh
#   /api/v1/users/*      — profiles, follow/unfollow
#   /api/v1/feed, /posts — timeline, create post, like, comment, stories
#   /api/v1/nearby       — geospatial discovery (query: lat,lng,radius,gender,age)
#   /api/v1/conversations, /messages, /ws — chat REST + WebSocket upgrade
#   /api/v1/media/uploads — presigned upload URLs